Add YAML config reading abstraction.
Currently the YAML config file is read in multiple places.
Adding this abstraction will more easily allow us to use it
in a single place, and not read the config for every lambda function.

It also provides an abstraction for each function in the config so we're
less reliant on hashes.