Isolate asset uploading implementation to the pusher class.
The intention here is that we use the config to determine which
'pusher' we need. And to remove the dependency in the lambda function
updater into the pusher class.

This will allow us to extend the solution, to have other pushers
i.e. an ECR Pusher. As well as a 'builder'.

The implementation here still assumes ruby is the runtime and
S3 is the asset repository. But it opens the door to enable others
in the future.