Support deployment of containerised R lambdas
This commit adds an R Builder and ECR Pusher - two classes required
to allow us to handle deployments for R lambda functions which are
defined in a Docker container. The `build` step requires the building
of a Docker image based on the Dockerfile specified in the function
config, `push` involves tagging the image with a short git sha and
uploading it to an ECR repository (also added to function config).
The `deploy` step, i.e. updating the function to use the latest version
of the image, uses the existing LambdaUpdater class but calls it with
the new image_uri instead of S3 object details.