Enables forcing a deployment without a new commit. (#13)
* Enables forcing a deployment without a new commit.

This behaviour is 'dangerous' as it overwrites the previous
deployments code. But is useful for development when iterating
where you may not want to commit your changes.

Also relies on the Thor CLI defaults to populate the overrides
to allow us to handle less.

* Refactor Function Deployer to orchestrate the deployment.

The Function Deployer will use the 'overrides' and state of the
builder, pusher and updater to determine which path to take.
This removes the responsibility of the implementations to check for
previous versions. Ultimately, it becomes simpler.

* Rename overrides to options

* Update ECR Pusher to not worry about checking if the assets are already in the repo

* Add log when assets aren't updated

* Tidy options argument in deploy method

* Update the version to 0.5.0