Raise errors fom system calls.
The deployment command relies heavily on calling out
to subprocesses for building and pushing assets. However,
we executed these processes in different ways and did not
raise exceptions if they failed. Therefore it was easy for
deployments to appear successful when they should have failed.

This isn't the entire story, as the updates to the Lambda function
are also caught to enable commenting on PRs. This will be addressed
seperately.