Set non-zero exit code when Thor errors.
By default any errors captured by Thor will output a
zero error code. This is not the behaviour we want as the
CLI is usually ran in Github so we want to fail the process.

Note this doesn't effect how we raise errors in our code, which
will raise non-zero exit codes regardless of this setting.