Raise errors fom system calls. (#113)
* Raise errors fom system calls.

The deployment command relies heavily on calling out
to subprocesses for building and pushing assets. However,
we executed these processes in different ways and did not
raise exceptions if they failed. Therefore it was easy for
deployments to appear successful when they should have failed.

This isn't the entire story, as the updates to the Lambda function
are also caught to enable commenting on PRs. This will be addressed
seperately.

* Set non-zero exit code when Thor errors.

By default any errors captured by Thor will output a
zero error code. This is not the behaviour we want as the
CLI is usually ran in Github so we want to fail the process.

Note this doesn't effect how we raise errors in our code, which
will raise non-zero exit codes regardless of this setting.

* Add slighly better error messages.

The error messages lack a little bit of infomation so I've added a
little bit more verbosity here.

For example, unless you know why don't update assets with a matching
git sha the Asserts not updated log probably looks like something has
gone wrong.

Furthermore, if a Lambda function fails to update we would like to know
why.

* Raise execption when unknown actions is specified

* Skip line length check for console output