Raise Lambda update errors from the function_deployer.
Will halt the execution of the Github Job which is desired in most
cases. Elevates the logging to Github to a higher level, where it
is better suited as it is an orchestration task, not a specific deployment
concern.

There's a slight change of functionality here and a bit of a braking
change for the staging deploy workflow. Users will be required to update
their workflow to continue staging deploys when the job fails, the job
will now also 'fail fast', i.e. if it fails on the first function it
will not attempt to deploy the next function. I think this is better as
we want to get to the problem as soon as possible.