Enable legacy compatibility for Minitest (for Mocha)
The Dockerfile uses an old version of Ruby (see #180). The latest version of Minitest removed some old compatibility code to support its old name (MiniTest), and Mocha was updated also to use its latest name. But the image is stuck with the last version of Mocha supported on 2.7, which still uses the old MiniTest name.

Setting this env var re-enables the legacy compatibility code for Minitest.