Add logging for when an image can't be found in ECR. (#199)
We have had occassions when deploying when an image can
not be found in ECR. However, when we deploy, we expect
the image to be there. There are a number of reasons
why this could be. We're adding the logging to test
what the response is from ECR, and what tag we're trying
to use. This should help narrow down the issue.