Add max_results to describe_images call.
We occassionally hit a bug where we have too many images
in our repository and on deploy we can't find the correct image tag as
it is not included on the first page of the results.

Our ECR repositories are configured to retain 100 images, but there is
latency in the deletion of the old images so we can hit this.

As the repos are limited to 100, the simplest fix is to bump the
max results to 1000.