Updating release guidance
# What
The gem release step expects to create a tag for the new gem version.
Manually doing the Release Notes and creating a tag before this action has completed will cause the release workflow to fail at the gem release step.
This causes the docker image step to be skipped, resulting in:
- dependabot being aware of a new release
- the gem version being available
- no docker imaage for the release being available.

Dev-P are a bit strapped for time just now, so I'm documenting our way around the issue for now.

Dev-P Ticket:
- https://github.com/fac/dev-platform/issues/2283